simplified the code into counterbutton component and add reset button
